misc: unify Image new-type
